Substitution

Substitute symbols in a freealg object for numbers or other freealg objects

subs(...)
subsu(S1, S2, r)

Arguments

S1, S2

Objects of class freealg

r

Integer specifying symbol to substitute (\(a=1,b=2\) etc)

...

First element is a freealg object, remainder are named arguments corresponding to variables to substitute

Value

Returns a freealg object.

Details

Function subs() substitutes variables for freealg objects (coerced if necessary) using natural R idiom. Observe that this type of substitution is sensitive to order:

> subs("ax",a="1+x",x="1+a")
free algebra element algebraically equal to
 + 2 + 3*a + 1*aa

> subs("ax",x="1+a",a="1+x")
free algebra element algebraically equal to
 + 2 + 3*x + 1*xx

Functions subsu() is a lower-level formal function, not really intended for the end-user. Function subsu() takes S1 and substitutes occurrences of symbol r with S2.

No equivalent to mvp::subvec() is currently implemented.

Author

Robin K. S. Hankin

Note

Function subs() is one place in the package where the use of letters is effectively hard-wired in. Idiom such as

subs("abccc",b="1+3x")

is very nice, but identifies “b” with 2. Note that argument r of subsu() is canonically an integer but a single character is interpreted as a letter. See also the note at freealg.Rd.

Examples

subs("abccc", b="1+3x")
#> free algebra element algebraically equal to
#> + accc + 3axccc
subs("aaaa", a="1+x")  # binomial
#> free algebra element algebraically equal to
#> + 1 + 4x + 6xx + 4xxx + xxxx

subs("abA", b=31)
#> free algebra element algebraically equal to
#> + 31

subs("1+a", a="A")   # can substitute for an inverse
#> free algebra element algebraically equal to
#> + 1 + A
subs("A", a="1+x")   # inverses are not substituted for
#> free algebra element algebraically equal to
#> + A

a <- rfalg()
subs(a, a=a)
#> free algebra element algebraically equal to
#> + 54aacaacaacac + 81aacaacaacacaa + 270aacaacaacb + 135aacaacaacc +
#> 189aacaacaaccbb + 27aacaacaacccb + 36aacaaccac + 54aacaaccacaa + 180aacaaccb +
#> 90aacaaccc + 126aacaacccbb + 18aacaaccccb + 180aacabcac + 270aacabcacaa +
#> 900aacabcb + 450aacabcc + 630aacabccbb + 90aacabcccb + 126aacacbbcac +
#> 189aacacbbcacaa + 630aacacbbcb + 315aacacbbcc + 441aacacbbccbb + 63aacacbbcccb
#> + 90aacaccac + 135aacaccacaa + 450aacaccb + 18aacaccbcac + 27aacaccbcacaa +
#> 90aacaccbcb + 45aacaccbcc + 63aacaccbccbb + 9aacaccbcccb + 225aacaccc +
#> 315aacacccbb + 45aacaccccb + 6acaac + 4acc + 10b + 180bacaacac + 270bacaacacaa
#> + 900bacaacb + 450bacaacc + 630bacaaccbb + 90bacaacccb + 120baccac +
#> 180baccacaa + 600baccb + 300baccc + 420bacccbb + 60baccccb + 600bbcac +
#> 900bbcacaa + 126bbcacaacac + 189bbcacaacacaa + 630bbcacaacb + 315bbcacaacc +
#> 441bbcacaaccbb + 63bbcacaacccb + 84bbcaccac + 126bbcaccacaa + 420bbcaccb +
#> 210bbcaccc + 294bbcacccbb + 42bbcaccccb + 3000bbcb + 420bbcbcac + 630bbcbcacaa
#> + 2100bbcbcb + 1050bbcbcc + 1470bbcbccbb + 210bbcbcccb + 1500bbcc + 2100bbccbb
#> + 294bbccbbcac + 441bbccbbcacaa + 1470bbccbbcb + 735bbccbbcc + 1029bbccbbccbb +
#> 147bbccbbcccb + 210bbcccac + 315bbcccacaa + 1350bbcccb + 42bbcccbcac +
#> 63bbcccbcacaa + 210bbcccbcb + 105bbcccbcc + 147bbcccbccbb + 21bbcccbcccb +
#> 525bbcccc + 735bbccccbb + 105bbcccccb + 20bc + 420bcbbcac + 630bcbbcacaa +
#> 2100bcbbcb + 1050bcbbcc + 1470bcbbccbb + 210bcbbcccb + 300bccac + 450bccacaa +
#> 18bccacaacac + 27bccacaacacaa + 90bccacaacb + 45bccacaacc + 63bccacaaccbb +
#> 9bccacaacccb + 12bccaccac + 18bccaccacaa + 60bccaccb + 30bccaccc + 42bccacccbb
#> + 6bccaccccb + 1500bccb + 120bccbcac + 180bccbcacaa + 600bccbcb + 300bccbcc +
#> 420bccbccbb + 60bccbcccb + 750bccc + 1050bcccbb + 42bcccbbcac + 63bcccbbcacaa +
#> 210bcccbbcb + 105bcccbbcc + 147bcccbbccbb + 21bcccbbcccb + 30bccccac +
#> 45bccccacaa + 300bccccb + 6bccccbcac + 9bccccbcacaa + 30bccccbcb + 15bccccbcc +
#> 21bccccbccbb + 3bccccbcccb + 75bccccc + 105bcccccbb + 15bccccccb + 5c +
#> 36caacaacac + 54caacaacacaa + 180caacaacb + 90caacaacc + 126caacaaccbb +
#> 18caacaacccb + 24caaccac + 36caaccacaa + 120caaccb + 60caaccc + 84caacccbb +
#> 12caaccccb + 120cabcac + 180cabcacaa + 600cabcb + 300cabcc + 420cabccbb +
#> 60cabcccb + 90cacaacac + 135cacaacacaa + 450cacaacb + 225cacaacc + 315cacaaccbb
#> + 45cacaacccb + 84cacbbcac + 126cacbbcacaa + 420cacbbcb + 210cacbbcc +
#> 294cacbbccbb + 42cacbbcccb + 120caccac + 180caccacaa + 600caccb + 12caccbcac +
#> 18caccbcacaa + 60caccbcb + 30caccbcc + 42caccbccbb + 6caccbcccb + 300caccc +
#> 420cacccbb + 60caccccb + 7cbb + 14cbbc + 300cbcac + 450cbcacaa + 1500cbcb +
#> 750cbcc + 1050cbccbb + 150cbcccb + 10cc + ccb + 210ccbbcac + 315ccbbcacaa +
#> 1050ccbbcb + 525ccbbcc + 735ccbbccbb + 105ccbbcccb + 2ccbc + 150cccac +
#> 225cccacaa + 750cccb + 30cccbcac + 45cccbcacaa + 150cccbcb + 75cccbcc +
#> 105cccbccbb + 15cccbcccb + 375cccc + 525ccccbb + 75cccccb

## Sequential substitution works:

subs("abccc", b="1+3x", x="1+d+2e")
#> free algebra element algebraically equal to
#> + 4accc + 3adccc + 6aeccc
subs(rfalg(), a=rfalg())
#> free algebra element algebraically equal to
#> + 7acaacbaacb + 28acaacbab + 14acaacbbaab + 21acaacbbbcb + 77acaacbc +
#> 49acaacbca + 28acabaacb + 112acabab + 56acabbaab + 84acabbbcb + 308acabc +
#> 196acabca + 14acbaabaacb + 56acbaabab + 28acbaabbaab + 42acbaabbbcb +
#> 154acbaabc + 98acbaabca + 21acbbcbaacb + 84acbbcbab + 42acbbcbbaab +
#> 63acbbcbbbcb + 231acbbcbc + 147acbbcbca + 49accaaacb + 196accaab + 77accaacb +
#> 308accab + 98accabaab + 147accabbcb + 539accac + 343accaca + 154accbaab +
#> 231accbbcb + 847accc + 539accca + 4baaacbaacb + 16baaacbab + 8baaacbbaab +
#> 12baaacbbbcb + 44baaacbc + 28baaacbca + 16baabaacb + 2baabaacbaacb +
#> 8baabaacbab + 4baabaacbbaab + 6baabaacbbbcb + 22baabaacbc + 14baabaacbca +
#> 64baabab + 8baababaacb + 32baababab + 16baababbaab + 24baababbbcb + 88baababc +
#> 56baababca + 32baabbaab + 4baabbaabaacb + 16baabbaabab + 8baabbaabbaab +
#> 12baabbaabbbcb + 44baabbaabc + 28baabbaabca + 48baabbbcb + 6baabbbcbaacb +
#> 24baabbbcbab + 12baabbbcbbaab + 18baabbbcbbbcb + 66baabbbcbc + 42baabbbcbca +
#> 176baabc + 112baabca + 14baabcaaacb + 56baabcaab + 22baabcaacb + 88baabcab +
#> 28baabcabaab + 42baabcabbcb + 154baabcac + 98baabcaca + 44baabcbaab +
#> 66baabcbbcb + 242baabcc + 154baabcca + 8babaabaacb + 32babaabab + 16babaabbaab
#> + 24babaabbbcb + 88babaabc + 56babaabca + 12babbcbaacb + 48babbcbab +
#> 24babbcbbaab + 36babbcbbbcb + 132babbcbc + 84babbcbca + 28bacaaacb + 112bacaab
#> + 44bacaacb + 176bacab + 56bacabaab + 84bacabbcb + 308bacac + 196bacaca +
#> 88bacbaab + 132bacbbcb + 484bacc + 308bacca + 7bb + bcaaaacbaacb + 4bcaaaacbab
#> + 2bcaaaacbbaab + 3bcaaaacbbbcb + 11bcaaaacbc + 7bcaaaacbca + 4bcaaabaacb +
#> 16bcaaabab + 8bcaaabbaab + 12bcaaabbbcb + 44bcaaabc + 28bcaaabca +
#> 2bcaabaabaacb + 8bcaabaabab + 4bcaabaabbaab + 6bcaabaabbbcb + 22bcaabaabc +
#> 14bcaabaabca + 3bcaabbcbaacb + 12bcaabbcbab + 6bcaabbcbbaab + 9bcaabbcbbbcb +
#> 33bcaabbcbc + 21bcaabbcbca + 7bcaacaaacb + 28bcaacaab + 11bcaacaacb + 44bcaacab
#> + 14bcaacabaab + 21bcaacabbcb + 77bcaacac + 49bcaacaca + 22bcaacbaab +
#> 33bcaacbbcb + 121bcaacc + 77bcaacca + 3bcbbaacbaacb + 12bcbbaacbab +
#> 6bcbbaacbbaab + 9bcbbaacbbbcb + 33bcbbaacbc + 21bcbbaacbca + 12bcbbabaacb +
#> 48bcbbabab + 24bcbbabbaab + 36bcbbabbbcb + 132bcbbabc + 84bcbbabca +
#> 6bcbbbaabaacb + 24bcbbbaabab + 12bcbbbaabbaab + 18bcbbbaabbbcb + 66bcbbbaabc +
#> 42bcbbbaabca + 9bcbbbbcbaacb + 36bcbbbbcbab + 18bcbbbbcbbaab + 27bcbbbbcbbbcb +
#> 99bcbbbbcbc + 63bcbbbbcbca + 21bcbbcaaacb + 84bcbbcaab + 33bcbbcaacb +
#> 132bcbbcab + 42bcbbcabaab + 63bcbbcabbcb + 231bcbbcac + 147bcbbcaca +
#> 66bcbbcbaab + 99bcbbcbbcb + 363bcbbcc + 231bcbbcca + 7bccb + 6c + 11caacbaacb +
#> 44caacbab + 22caacbbaab + 33caacbbbcb + 121caacbc + 77caacbca + 44cabaacb +
#> 176cabab + 88cabbaab + 132cabbbcb + 484cabc + 308cabca + 22cbaabaacb +
#> 88cbaabab + 44cbaabbaab + 66cbaabbbcb + 242cbaabc + 154cbaabca + 4cbaacb +
#> 16cbab + 21cbacaacb + 84cbacab + 42cbacbaab + 63cbacbbcb + 231cbacc + 147cbacca
#> + 12cbbaaacb + 56cbbaab + 6cbbaabaacb + 24cbbaabab + 12cbbaabbaab +
#> 18cbbaabbbcb + 66cbbaabc + 42cbbaabca + 24cbbabaab + 36cbbabbcb + 132cbbac +
#> 84cbbaca + 12cbbbcb + 3cbbcaaaacb + 12cbbcaaab + 6cbbcaabaab + 9cbbcaabbcb +
#> 33cbbcaac + 21cbbcaaca + 33cbbcbaacb + 132cbbcbab + 66cbbcbbaab + 9cbbcbbaacb +
#> 36cbbcbbab + 18cbbcbbbaab + 27cbbcbbbbcb + 99cbbcbbbcb + 99cbbcbbc + 63cbbcbbca
#> + 363cbbcbc + 231cbbcbca + 44cbc + 28cbca + 33cbcaacb + 132cbcab + 66cbcbaab +
#> 99cbcbbcb + 363cbcc + 231cbcca + 77ccaaacb + 308ccaab + 121ccaacb + 484ccab +
#> 154ccabaab + 231ccabbcb + 847ccac + 539ccaca + 242ccbaab + 363ccbbcb + 1331ccc
#> + 847ccca