Single-letter symbols

Variables a, b,..., z and their inverses A-Z are given their freealg semantic meaning.

Details

Sometimes it is convenient in an R session to have all 26 letters a-z and all 26 uppercase letters A-Z adopt their free algebra interpretations. To access this, load the lettersymbols dataset, which is provided with the package in the inst directory:

  load(system.file("lettersymbols.rda",package="freealg"))

Executing this allows you to do cool things such as the following:

> (1+a-b^2)^4
free algebra element algebraically equal to
+ 1 + 4a + 6aa + 4aaa + aaaa - aaabb - 4aabb - aabba + aabbbb - 6abb - 4abba -
abbaa + abbabb + 4abbbb + abbbba - abbbbbb - 4bb - 6bba - 4bbaa - bbaaa +
bbaabb + 4bbabb + bbabba - bbabbbb + 6bbbb + 4bbbba + bbbbaa - bbbbabb -
4bbbbbb - bbbbbba + bbbbbbbb
>

Lowercase letters c, q, t, and uppercase letters C, D, F, I, T might pose difficulties.

These objects can also be generated by running script inst/symb.Rmd, which includes some further discussion and technical documentation and creates file lettersymbols.rda which formerly resided in the data/ directory.

Author

Robin K. S. Hankin