Substitution

Substitute symbols in a freealg object for numbers or other freealg objects

Usage

subs(S, ...)
subsu(S1,S2,r)

Arguments

S,S1,S2

Objects of class freealg

r

Integer specifying symbol to substitute (\(a=1,b=2\) etc)

...

named arguments corresponding to variables to substitute

Value

Returns a freealg object.

Details

Function subs() substitutes variables for freealg objects (coerced if necessary) using natural R idiom. Observe that this type of substitution is sensitive to order:

> subs("ax",a="1+x",x="1+a")
free algebra element algebraically equal to
 + 2 + 3*a + 1*aa

> subs("ax",x="1+a",a="1+x")
free algebra element algebraically equal to
 + 2 + 3*x + 1*xx

Functions subsu() is a lower-level formal function, not really intended for the end-user. Function subsu() takes S1 and substitutes occurrences of symbol r with S2.

No equivalent to mvp::subvec() is currently implemented.

Author

Robin K. S. Hankin

Note

Function subs() is one place in the package where the use of letters is effectively hard-wired in. Idiom such as

subs("abccc",b="1+3x")

is very nice, but identifies “b” with 2. Note that argument r of subsu() is canonically an integer but a single character is interpreted as a letter. See also the note at freealg.Rd.

Examples

subs("abccc",b="1+3x")
#> free algebra element algebraically equal to
#> + accc + 3axccc
subs("aaaa",a="1+x")  # binomial
#> free algebra element algebraically equal to
#> + 1 + 4x + 6xx + 4xxx + xxxx

subs("abA",b=31)
#> free algebra element algebraically equal to
#> + 31

subs("1+a",a="A")   # can substitute for an inverse
#> free algebra element algebraically equal to
#> + 1 + A
subs("A",a="1+x")   # inverses are not substituted for
#> free algebra element algebraically equal to
#> + A

## Sequential substitution works:

subs("abccc",b="1+3x",x="1+d+2e")
#> free algebra element algebraically equal to
#> + 4accc + 3adccc + 6aeccc
subs(rfalg(),a=rfalg())
#> free algebra element algebraically equal to
#> + 1512aaa + 1764aabc + 756aaca + 1008aacaba + 252aacabb + 1764aaccb + 1764abca
#> + 2058abcbc + 882abcca + 1176abccaba + 294abccabb + 2058abcccb + 6ac + 1188acaa
#> + 1008acabaa + 1176acababc + 504acabaca + 672acabacaba + 168acabacabb +
#> 1176acabaccb + 252acabba + 294acabbbc + 126acabbca + 168acabbcaba + 42acabbcabb
#> + 294acabbccb + 1386acabc + 594acaca + 792acacaba + 198acacabb + 1386acaccb +
#> 504acbca + 588acbcbc + 252acbcca + 336acbccaba + 84acbccabb + 588acbcccb +
#> 216accaa + 288accabaa + 336accababc + 144accabaca + 192accabacaba +
#> 48accabacabb + 336accabaccb + 72accabba + 84accabbbc + 36accabbca +
#> 48accabbcaba + 12accabbcabb + 84accabbccb + 252accabc + 108accaca + 144accacaba
#> + 36accacabb + 252accaccb + 1764accba + 2058accbbc + 882accbca + 1176accbcaba +
#> 294accbcabb + 2058accbccb + 504acccba + 588acccbbc + 252acccbca + 336acccbcaba
#> + 84acccbcabb + 588acccbccb + 8b + 1764bcaa + 2058bcabc + 882bcaca +
#> 1176bcacaba + 294bcacabb + 2058bcaccb + 2058bcbca + 2401bcbcbc + 1029bcbcca +
#> 1372bcbccaba + 343bcbccabb + 2401bcbcccb + 7bcc + 1386bccaa + 1176bccabaa +
#> 1372bccababc + 588bccabaca + 784bccabacaba + 196bccabacabb + 1372bccabaccb +
#> 294bccabba + 343bccabbbc + 147bccabbca + 196bccabbcaba + 49bccabbcabb +
#> 343bccabbccb + 1617bccabc + 693bccaca + 924bccacaba + 231bccacabb + 1617bccaccb
#> + 588bccbca + 686bccbcbc + 294bccbcca + 392bccbccaba + 98bccbccabb +
#> 686bccbcccb + 252bcccaa + 336bcccabaa + 392bcccababc + 168bcccabaca +
#> 224bcccabacaba + 56bcccabacabb + 392bcccabaccb + 84bcccabba + 98bcccabbbc +
#> 42bcccabbca + 56bcccabbcaba + 14bcccabbcabb + 98bcccabbccb + 294bcccabc +
#> 126bcccaca + 168bcccacaba + 42bcccacabb + 294bcccaccb + 2058bcccba +
#> 2401bcccbbc + 1029bcccbca + 1372bcccbcaba + 343bcccbcabb + 2401bcccbccb +
#> 588bccccba + 686bccccbbc + 294bccccbca + 392bccccbcaba + 98bccccbcabb +
#> 686bccccbccb + 4c + 756caaa + 882caabc + 378caaca + 504caacaba + 126caacabb +
#> 882caaccb + 1008cabaaa + 1176cabaabc + 504cabaaca + 672cabaacaba + 168cabaacabb
#> + 1176cabaaccb + 1176cababca + 1372cababcbc + 588cababcca + 784cababccaba +
#> 196cababccabb + 1372cababcccb + 4cabac + 792cabacaa + 672cabacabaa +
#> 784cabacababc + 336cabacabaca + 448cabacabacaba + 112cabacabacabb +
#> 784cabacabaccb + 168cabacabba + 196cabacabbbc + 84cabacabbca + 112cabacabbcaba
#> + 28cabacabbcabb + 196cabacabbccb + 924cabacabc + 396cabacaca + 528cabacacaba +
#> 132cabacacabb + 924cabacaccb + 336cabacbca + 392cabacbcbc + 168cabacbcca +
#> 224cabacbccaba + 56cabacbccabb + 392cabacbcccb + 144cabaccaa + 192cabaccabaa +
#> 224cabaccababc + 96cabaccabaca + 128cabaccabacaba + 32cabaccabacabb +
#> 224cabaccabaccb + 48cabaccabba + 56cabaccabbbc + 24cabaccabbca +
#> 32cabaccabbcaba + 8cabaccabbcabb + 56cabaccabbccb + 168cabaccabc + 72cabaccaca
#> + 96cabaccacaba + 24cabaccacabb + 168cabaccaccb + 1176cabaccba + 1372cabaccbbc
#> + 588cabaccbca + 784cabaccbcaba + 196cabaccbcabb + 1372cabaccbccb +
#> 336cabacccba + 392cabacccbbc + 168cabacccbca + 224cabacccbcaba + 56cabacccbcabb
#> + 392cabacccbccb + 252cabbaa + 294cabbabc + 126cabbaca + 168cabbacaba +
#> 42cabbacabb + 294cabbaccb + 294cabbbca + 343cabbbcbc + 147cabbbcca +
#> 196cabbbccaba + 49cabbbccabb + 343cabbbcccb + cabbc + 198cabbcaa + 168cabbcabaa
#> + 196cabbcababc + 84cabbcabaca + 112cabbcabacaba + 28cabbcabacabb +
#> 196cabbcabaccb + 42cabbcabba + 49cabbcabbbc + 21cabbcabbca + 28cabbcabbcaba +
#> 7cabbcabbcabb + 49cabbcabbccb + 231cabbcabc + 99cabbcaca + 132cabbcacaba +
#> 33cabbcacabb + 231cabbcaccb + 84cabbcbca + 98cabbcbcbc + 42cabbcbcca +
#> 56cabbcbccaba + 14cabbcbccabb + 98cabbcbcccb + 36cabbccaa + 48cabbccabaa +
#> 56cabbccababc + 24cabbccabaca + 32cabbccabacaba + 8cabbccabacabb +
#> 56cabbccabaccb + 12cabbccabba + 14cabbccabbbc + 6cabbccabbca + 8cabbccabbcaba +
#> 2cabbccabbcabb + 14cabbccabbccb + 42cabbccabc + 18cabbccaca + 24cabbccacaba +
#> 6cabbccacabb + 42cabbccaccb + 294cabbccba + 343cabbccbbc + 147cabbccbca +
#> 196cabbccbcaba + 49cabbccbcabb + 343cabbccbccb + 84cabbcccba + 98cabbcccbbc +
#> 42cabbcccbca + 56cabbcccbcaba + 14cabbcccbcabb + 98cabbcccbccb + 882cabca +
#> 1029cabcbc + 441cabcca + 588cabccaba + 147cabccabb + 1029cabcccb + 3cac +
#> 594cacaa + 504cacabaa + 588cacababc + 252cacabaca + 336cacabacaba +
#> 84cacabacabb + 588cacabaccb + 126cacabba + 147cacabbbc + 63cacabbca +
#> 84cacabbcaba + 21cacabbcabb + 147cacabbccb + 693cacabc + 297cacaca +
#> 396cacacaba + 99cacacabb + 693cacaccb + 252cacbca + 294cacbcbc + 126cacbcca +
#> 168cacbccaba + 42cacbccabb + 294cacbcccb + 108caccaa + 144caccabaa +
#> 168caccababc + 72caccabaca + 96caccabacaba + 24caccabacabb + 168caccabaccb +
#> 36caccabba + 42caccabbbc + 18caccabbca + 24caccabbcaba + 6caccabbcabb +
#> 42caccabbccb + 126caccabc + 54caccaca + 72caccacaba + 18caccacabb + 126caccaccb
#> + 882caccba + 1029caccbbc + 441caccbca + 588caccbcaba + 147caccbcabb +
#> 1029caccbccb + 252cacccba + 294cacccbbc + 126cacccbca + 168cacccbcaba +
#> 42cacccbcabb + 294cacccbccb + 6cbb + 1764ccbaa + 2058ccbabc + 882ccbaca +
#> 1176ccbacaba + 294ccbacabb + 2058ccbaccb + 2058ccbbca + 2401ccbbcbc +
#> 1029ccbbcca + 1372ccbbccaba + 343ccbbccabb + 2401ccbbcccb + 7ccbc + 1386ccbcaa
#> + 1176ccbcabaa + 1372ccbcababc + 588ccbcabaca + 784ccbcabacaba + 196ccbcabacabb
#> + 1372ccbcabaccb + 294ccbcabba + 343ccbcabbbc + 147ccbcabbca + 196ccbcabbcaba +
#> 49ccbcabbcabb + 343ccbcabbccb + 1617ccbcabc + 693ccbcaca + 924ccbcacaba +
#> 231ccbcacabb + 1617ccbcaccb + 588ccbcbca + 686ccbcbcbc + 294ccbcbcca +
#> 392ccbcbccaba + 98ccbcbccabb + 686ccbcbcccb + 252ccbccaa + 336ccbccabaa +
#> 392ccbccababc + 168ccbccabaca + 224ccbccabacaba + 56ccbccabacabb +
#> 392ccbccabaccb + 84ccbccabba + 98ccbccabbbc + 42ccbccabbca + 56ccbccabbcaba +
#> 14ccbccabbcabb + 98ccbccabbccb + 294ccbccabc + 126ccbccaca + 168ccbccacaba +
#> 42ccbccacabb + 294ccbccaccb + 2058ccbccba + 2401ccbccbbc + 1029ccbccbca +
#> 1372ccbccbcaba + 343ccbccbcabb + 2401ccbccbccb + 588ccbcccba + 686ccbcccbbc +
#> 294ccbcccbca + 392ccbcccbcaba + 98ccbcccbcabb + 686ccbcccbccb